requirements.txt
alembic==1.4.3
appdirs==1.4.4
astroid==2.4.1
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.2
awscli==1.18.190
Babel==2.8.0
botocore==1.19.30
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
distlib==0.3.1
docutils==0.15.2
ecdsa==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
iniconfig==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.7
pluggy==0.13.1
postgres==3.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
py==1.9.0
pyasn1==0.4.8
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.5.2
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==1.5
python-dotenv==0.15.0
python-editor==1.0.4
python-jose==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.25.1
rsa==4.5
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.3.19
toml==0.10.0
urllib3==1.26.2
virtualenv==20.0.33
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
